refactor: Better webring icon infrastructure
