+++ components/common/webring/LionSvg.tsx
